requirements.txt
Flask==3.0.3
flask-cors==4.0.0
azure-ai-inference>=1.0.0
azure-core>=1.30.0
urllib3<2
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
